pyproject.toml
[tool.poetry]
name = "flask_ease"
version = "0.1.2"
description = ""
authors = ["Shubham Biswas <shubhambiswas.zero@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
flask = "^1.1.2"
pydantic = "^1.5.1"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
flask-sqlalchemy = "^2.4.3"
alembic = "^1.4.2"
autopep8 = "^1.5.3"
pydantic = "^1.5.1"
python-dotenv = "^0.14.0"
psycopg2-binary = "^2.8.5"
flask_cors = "^3.0.8"
passlib = "^1.7.2"
bcrypt = "^3.1.7"
pytz = "^2020.1"
pyjwt = "^1.7.1"
twine = "^3.2.0"
setuptools = "^49.1.1"
wheel = "^0.34.2"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
